behavior change -> New Window opens with same URL
This patch will cause a "New Window" command to open a new window with
the URL from the active view.  Opening a new instance via launcher will
still open a new window with the URL specified in Startup settings.
This is the behavior in Windows 10 file explorer.  The advantage is this
gives the user a way to open a new window with the current URL(New
Window) as well as a new window with URL specified in Startup
settings(new instance) whereas currently, there isn't an easy way to
open a new window with the current URL.  You'd have to either open a new
window and copy and paste the URL, or open a new tab(which opens with
the same URL) and detach it.  I'd much rather have the new window open
with the current URL and if what I wanted was the Startup URL I could
easily get there as it's already bookmarked.

REVIEW: 129973