Sort the tag-values alphabetically in the "Tags" column
Up until now tag-values have appeared unsorted in the Tags column
when the selected View Mode is "Details".
In older versions of Dolphin (in KDE4) the tags-values were
alphabetically sorted in the Tags column, which means that back then
this was the desired behavior.
This commit restores this functionality.

BUG: 377589
FIXED-IN: 17.04.1
REVIEW: 130068