Don't ignore tag clicks in the tooltips
Summary:
Now that we can use the metadata widgets in the tooltips, we can also
open the tags:// url if the user clicks some tag in a tooltip.

The behavior is now consistent with the metadata widget in the information panel.

Test Plan: Click a tag when the metadata tooltip shows up.

Reviewers: emmanuelp

Subscribers: #konqueror, #dolphin

Differential Revision: https://phabricator.kde.org/D5658