Make sure we always have Shift+Del as shortcut
After commit 68bb0ec22a the shortcut for the Delete action is not
necessarily Shift+Del, but whatever the user set in System Setting.
However DolphinRemoveAction assumes/hardcodes Shift+Del, so we should
always make sure we have this shortcut around, for consistency.

We just need to add it (if necessary) to the list of shortcuts of the
action. However:

* for the actual Delete action, we need to append it (if we'd prepend it,
  it would override a custom primary shortcut in the 'Configure Shortcuts' dialog).
* for DolphinRemoveAction, we need to prepend it in order to have
  Shift+Del (rather than the custom primary shortcut) in the context menu.

Differential Revision: https://phabricator.kde.org/D7655