Fix build with Qt 5.10
Summary: QStaticText's constructor with const QString & as argument is now explicit.

Test Plan:
Did not build before, now does.

Note that this might change the behaviour as the other properties of m_text are now
kept, which was not the case before. I'm not sure what the expected behaviour here
is supposed to be.

Reviewers: #dolphin, #kde_applications, cfeck

Reviewed By: cfeck

Subscribers: cfeck, #kde_applications, #dolphin

Differential Revision: https://phabricator.kde.org/D8075