Make Saved Search feature discoverable
Summary:
FEATURE: 269332

Make Dolphin's Saved Search feature discoverable by adding a button inside the search field. The button only becomes enabled when there is a valid search term. When pushed, it saves the search to the Places panel, providing a visible-by-default way to do this to complement the existing implementation that is only visible in the context menu.

Also harmonized the label text so that it's consistent no matter how you create a saved search (button or context menu)

Test Plan:
Tested in KDE Neon. Works great:
{F5449508}

Reviewers: #dolphin, broulik, dfaure, markg, emateli, elvisangelaccio

Reviewed By: #dolphin, markg, emateli, elvisangelaccio

Subscribers: anthonyfieroni, markg, emateli, elvisangelaccio, cfeck, #dolphin

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D8454