Turn on Dolphin icon previews by default
Summary:
BUG: 338492
BUG: 350212

By default, turn on all preview plugins and turn on previews themselves.

Depends on D8347

Test Plan:
Tested this change in an up-to-date KDE Neon: removed the existing dolphinrc file, deployed Dolphin with the change, and observed that previews are now turned on for all file types for which a plugin exists except for text files, which are in the blacklist because they're mostly useless at nearly all icon sizes (still available in case people want them, though).

Here's how Dolphin's main window looks by default now in KDE Neon:

{F5441184}
(You may notice that there are no previews for the video files; that's because Neon doesn't ship with any plugins for them, so that would be expected at this point)

And here is how the Settings > General > Preview window looks like now by default:
{F5435391}

Reviewers: #dolphin, #vdg, markg, abetts, dfaure, aseigo, elvisangelaccio

Reviewed By: #vdg, abetts, dfaure, elvisangelaccio

Subscribers: abetts, cfeck, andreaska, emmanuelp, andreask, markg, broulik, anthonyfieroni, davidedmundson, ltoscano, dfaure, elvisangelaccio, #konqueror

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D7440