Elide the label text on the statusbar in the middle
Summary:
BUG: 387189

Currently the label text on the statusbar is elided in on the right, but that means that when the text gets elided the size of the file isn't shown.

The label text on the statusbar shows the file name, the mime type and the size; the name is already shown in the main window and the mimetype can usually be guessed from the icon or from the extension, so it's usually the size that the user wants to see on the statusbar.

Reviewers: #dolphin, ngraham, elvisangelaccio

Reviewed By: #dolphin, ngraham, elvisangelaccio

Subscribers: elvisangelaccio, ngraham, anthonyfieroni

Differential Revision: https://phabricator.kde.org/D8927