Select and scroll to the file for "Open path" and "Open Path in New Window" operations
Summary:
After performing a search and using  the "Open path" or "Open Path in New Window"  present in the context menu, make sure the file is selected and visible in the resulting view.

Cannot implement the same fix for "Open Path in New Tab" because of a limitation in how the contents of inactive tabs are rendered; will need to fix that separately in another patch.

BUG: 377510

Test Plan: Search for an item, and choose {nav Open Path} or {nav Open path in new window}

Reviewers: #dolphin, elvisangelaccio

Reviewed By: #dolphin, elvisangelaccio

Subscribers: elvisangelaccio, broulik

Differential Revision: https://phabricator.kde.org/D11703